Merge GoodPawn and EvilPawn classes.
If a new type of pawn was added, it would have had to choose between the two
classes to subclass, and thus for every type of pawn two more classes would have
been added.
"evil" and "good" are now stored as a "state" in a boolean in the Pawn class,
which is no more abstract.